change NUMOCCURRENCES from double to integer
git-svn-id: https://svn.code.sf.net/p/moneymanagerex/code/trunk@4735 7f5d68d0-c712-0410-839f-9b116a576cc5

+++ sql/tables_v1.sql
@@ -46,7 +46,7 @@ BDID integer primary key
 , TOTRANSAMOUNT numeric
 , REPEATS integer
 , NEXTOCCURRENCEDATE TEXT
-, NUMOCCURRENCES numeric
+, NUMOCCURRENCES integer 
 );
 
 -- Describe BUDGETSPLITTRANSACTIONS_V1
